Excised sourcereg file for projects with "keephistory=on"
Shouldn't affect projects with "keephistory=off".

Dual affects are loss of a lot of superfluous code and acre's ability to recognise and create a .delete file for an item deleted by the deletion of a sourcefile.

Will make the process of renaming a source folder by addition (or removal) of a casecode simpler.