Acre and SALT interop
If an edited item is not in a tracked project acre now passes control to
SALT to see if it can handle the item. The only difference that will be seen
is that no "No tracked project can handle item" message will be issued on
exiting the editor. SALT may issue its own messages.