Problems uncovered at Belfast
CaseCode and KeepHistory now default to 'Off' for a new project.
Call CreateProject with the -casecode='on' flag to reverse that.
Either can still be edited in the config file and will take effect at the
following OpenProject.
CreateProject will now pop up windows giving the names and ⎕NC of all items
NOT saved in the project whether by design or error.
APLAN is now capable of encoding and decoding arrays containing ⎕NULLs
'though not ⎕ORs or ⎕UCS 0.
But now prohibits the evaluation of expressions containing unquoted
characters other than '[(←⍴,⊂⋄/⍬⎕)]', ⎕D and name chars.