Defaults,Message Protocol, Reporting. Function access. Interop.
New projects are now created with CaseCode and KeepHistory 'Off'. Use the "-c=on" or "-k=on" flags to reverse them or they can then be edited in the config file. Its default name has changed from "acreconfig.txt" to "acre.config" and the default encoding is now APLAN.
All messages now appear, not in the session or a nominated array, but in a separate window which can be hidden, restored, cleared and queried. Now reports all items NOT written by CreateProject and any items that cannot be written if casecoding is Off because they would produce duplicate files.
Items not written by CreateProject and SetChanged are now listed.
There is an experimental protocol:
    ⎕SE.acreAPI.casecode
                changefile
                changes
                deletefile
                fromaplan
                itemname
                log
                nameclass
                projects
                sourcefile
                stats
                toaplan
                version
that is a namespace containing functions that reproduce all the facility
of ⎕SE.acre.run and a growing list of acre's user commands.
Acre now hands over to SALT if an edited item is not in any tracked project.