v14.0 emulators
Many small changes moving towards compatibility with Dyalog v14.0. All
v15.0 and v16.0 enhancements now used under control of "SV.aplver" set at
start, with older workarounds for previous vss. and new under-the-covers
function to build installation dwss into same. As and when acre is changed
to use further enhancements emulators will be written also for them.
At startup any emulator that is redundant because the present version
already contains that enhancement is overwritten with a much lighter cover
for the enhancement. For instance all calls to `⎕NGET` are covered by
7-line dfn `_NGET` that is overwritten with the read-only one-liner
`⊢{⊃'UTF-8'⎕NGET ⍵ 1}` in v15.0 and above and `@` is covered by 5-line `a_`
that is overwritten with the primitive itself in c16.0 and above.
So far covers for: ⎕MKDIR. ⎕NDELETE, ⎕NEXISTS, ⎕NGET, ⎕NINFO, ⎕NPARTS,
⎕NPUT, 819⌶, @, ⊆ & ⍸ but no attempt to make them anything other than
Windows compatible and perhaps there is no reason to.
They are not general solutions but cover only the specific cases required
by acre's present functionality.

New command `]close #.projectspace` stops tracking, leaves code in place
and returns the number of tracked projects remaining.