Acre 5.0.11+167 2019-04-09 22.39
Acre 5.0.11+167 2019-04-09 22.39

## Acre Desktop - Release Notes

As an attempt to implement Enhancement: #166 _Change of item-type in editor. e.g. Fn→Op_, a subfunction - called by the editor callback and the commands:
`]SetChanged`, `]Undo`, `]Redo` and `]Restore` - now checks to see if the _type_ of any item affected has changed and renames the source-file accordingly.

Causes test Misc_010
```
Change a function to an operator.
This must fail due to the fact that they have different file extensions
```
to fail because `]openProject` doesn't.