Acre 5.2.0+197 2019-06-17 10.34
Acre 5.2.0+197 2019-06-17 10.34

## Acre Desktop - Release Notes

Command `]Refresh` now added also as an API function `⎕SE.acreAPI.refresh`.

`]CreateProject` will now message also to session, in addition to log, if differently cased homonyms are not written due to default `-casecode=off` in Windows &c..
```
Can`t write duplicate files
===== ===== ========= =====
 #.proj.item  #.proj.ITEM
```

Fix to bug #168 applied also to `changes` and `refresh` to avoid the SYNTAX ERROR if the argument is not an open project or member thereof.